% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_AggrDat.R
\name{f_aggrDat}
\alias{f_aggrDat}
\title{Aggregate a local data frame and returning multiple summary statistics}
\usage{
f_aggrDat(dataframe, groupVars, valueVar, WideToLong = FALSE)
}
\arguments{
\item{dataframe}{dataframe}

\item{groupVars}{grouping variables}

\item{valueVar}{variable to aggregate}

\item{WideToLong}{change format of results dataset from wide to long format}
}
\description{
aggregates a single variable of a dataframe and returns summary statistics including CI's
}
