% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_AggrDat.R
\name{f_weighted.aggrDat}
\alias{f_weighted.aggrDat}
\title{dataframe = dataframe to aggregate (new datafram will be created)
groupVars = variables to aggregate at 
valueVar = variable to aggregate 
weightVar = weighting variable  
WideToLong = transfrom data to long format, so that statistics are in one column instead of spread over rows}
\usage{
f_weighted.aggrDat(
  dataframe,
  groupVars,
  valueVar,
  weightVar,
  WideToLong = FALSE
)
}
\description{
dataframe = dataframe to aggregate (new datafram will be created)
groupVars = variables to aggregate at 
valueVar = variable to aggregate 
weightVar = weighting variable  
WideToLong = transfrom data to long format, so that statistics are in one column instead of spread over rows
}
